refactor: rename schemes to colorscales
